requirements.txt
requests==2.26.0
python-dotenv==0.19.2
beautifulsoup4==4.10.0
langchain
ibm-watsonx-ai

Create an environment using venv

1. Open a terminal and navigate to your project folder.

python -m venv .venv

2. Any time you want to use your new environment, you first need to go to your project folder (where the .venv directory lives) and run the command to activate it:
# Windows command prompt
.venv\Scripts\activate.bat

# Windows PowerShell
.venv\Scripts\Activate.ps1

# macOS and Linux
source .venv/bin/activate

3. When you're done using this environment, return to your normal shell by typing:

deactivate